Completed exercise 4 of 13 -- learnyounode
Write a program that uses a single asynchronous filesystem operation to
read a f
ile and print the number of newlines it contains to the console
(stdout), simila
r to running cat file | wc -l.

The full path to the file to read will be provided as the first
command-line arg
ument.