Completed exercise 5 of 13 -- learnyounode
Create a program that prints a list of files in a given directory,
filtered by t
he extension of the files. You will be provided a directory name as the
first ar
gument to your program (e.g. '/path/to/dir/') and a file extension to
filter by
as the second argument.

For example, if you get 'txt' as the second argument then you will need
to filte
r the list to only files that end with .txt.

The list of files should be printed to the console, one file per line.
You must
use asynchronous I/O.