Completed 8 of 13 -- learnyounode
Write a program that performs an HTTP GET request to a URL provided to
you as th
e first command-line argument. Collect all data from the server (not
just the fi
rst "data" event) and then write two lines to the console (stdout).

The first line you write should just be an integer representing the
number of ch
aracters received from the server and the second line should contain the
complet
e String of characters sent by the server.