Completed 11 of 13 -- learnyounode
Write an HTTP server that serves the same text file for each request it
receives
.

Your server should listen on the port provided by the first argument to
your pro
gram.

You will be provided with the location of the file to serve as the
second comman
d-line argument. You must use the fs.createReadStream() method to stream
the fil
e contents to the response.